Add files via upload
The total of sales figures stands at 15.26M, with the higher share coming from the Electronics and Appliances categories. The sales distribution by customer age bins shows that the major purchases are done by the 30 to 50 age group, thus indicating a target audience for marketing campaigns. The trends from month to month can be analyzed and assessed from the sales numbers, which seem to spike rather sharply in some months before plummeting. This feature enables us to have some depth in understanding seasonal variation and probably the impact of promotions. The summary of daily sales incorporates further corroborations of the variability, with the most conspicuous decline coming during the final part of the period being analyzed. Weekday sales illustrated what would amount to a relatively even distribution across weekdays, indicating that the purchases were fairly determined by the day of the week. With reference to the map view of the sales per city and stores, there is a distinguished performance with the urban agglomerations of Los Angeles, New York, Miami, and Chicago. Another beneficial observation is that the level of inventory is positively correlated with total sales, as excess inventory increases sales. The demand trend applied to forecast sales is wavy, indicating to a higher or lower demand pattern influenced by external stimuli such as promotions, seasonality, or the prevailing state of the economy.

In summary, there is an opportunity for inventory optimization, promotions targeting high-sale customer segments, and more intense research into seasonal sales to improve the business strategy.